requirements/requirements.txt
amqp==2.5.0
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.0.2
django-import-export==1.2.0
django-timezone-field==3.0
djangorestframework==3.10.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
kombu==4.6.3
odfpy==1.4.0
openpyxl==2.6.2
postmarker==0.13.0
psycopg2==2.8.3
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
PyYAML==5.1.2
redis==3.3.6
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
